+++ yandex/resource_yandex_ydb_database_iam_binding.go
@@ -1,0 +1,7 @@
+package yandex
+
+import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+
+func resourceYandexYDBDatabaseIAMBinding() *schema.Resource {
+	return resourceIamBindingWithImport(IamYDBDatabaseSchema, newYDBDatabaseIamUpdater, ydbDatabaseIDParseFunc)
+}