+++ CHANGELOG.md
@@ -16,6 +16,7 @@ FEATURES:
 * k8s: add `instance_template.ipv4_dns_records`, `instance_template.ipv6_dns_records` attributes to `node group` resource and data source
 * greenplum: add fields `greenplum_config` and `pooler_config` in `resource_yandex_mdb_greenplum_cluster`
 * **New Resource:** `yandex_serverless_container_iam_binding`
+* **New Resource:** `yandex_ydb_database_iam_binding`
 
 ## 0.76.0 (July 01, 2022)
 BUG FIXES: